+++ src/cmd_rgr/mod.rs
@@ -2,3 +2,4 @@
 
 pub mod merge;
 pub mod replace;
+pub mod runlist;