+++ src/cmd_spanr/mod.rs
@@ -9,7 +9,6 @@ pub mod coverage;
 pub mod genome;
 pub mod gff;
 pub mod merge;
-pub mod range;
 pub mod some;
 pub mod span;
 pub mod split;