+++ tests/cli_rgr.rs
@@ -66,3 +66,86 @@ fn command_replace_reverse() -> Result<(), Box<dyn std::error::Error>> {
 
     Ok(())
 }
+
+#[test]
+fn command_runlist() -> Result<(), Box<dyn std::error::Error>> {
+    let mut cmd = Command::cargo_bin("rgr")?;
+    let output = cmd
+        .arg("runlist")
+        .arg("tests/rgr/intergenic.yml")
+        .arg("tests/rgr/S288c.rg")
+        .arg("--op")
+        .arg("overlap")
+        .output()
+        .unwrap();
+    let stdout = String::from_utf8(output.stdout).unwrap();
+
+    assert_eq!(stdout.lines().count(), 2);
+    assert!(!stdout.contains("S288c"));
+    assert!(stdout.contains("21294-22075"));
+
+    let mut cmd = Command::cargo_bin("rgr")?;
+    let output = cmd
+        .arg("runlist")
+        .arg("tests/rgr/intergenic.yml")
+        .arg("tests/rgr/S288c.rg")
+        .arg("--op")
+        .arg("non-overlap")
+        .output()
+        .unwrap();
+    let stdout = String::from_utf8(output.stdout).unwrap();
+
+    assert_eq!(stdout.lines().count(), 4);
+    assert!(stdout.contains("S288c"));
+    assert!(!stdout.contains("21294-22075"));
+
+    let mut cmd = Command::cargo_bin("rgr")?;
+    let output = cmd
+        .arg("runlist")
+        .arg("tests/rgr/intergenic.yml")
+        .arg("tests/rgr/S288c.rg")
+        .arg("--op")
+        .arg("superset")
+        .output()
+        .unwrap();
+    let stdout = String::from_utf8(output.stdout).unwrap();
+
+    assert_eq!(stdout.lines().count(), 2);
+    assert!(!stdout.contains("S288c"));
+    assert!(stdout.contains("21294-22075"));
+
+    Ok(())
+}
+
+#[test]
+fn command_runlist_invalid() -> Result<(), Box<dyn std::error::Error>> {
+    let mut cmd = Command::cargo_bin("rgr")?;
+    cmd.arg("runlist")
+        .arg("tests/rgr/intergenic.yml")
+        .arg("tests/rgr/S288c.rg")
+        .arg("--op")
+        .arg("invalid");
+    cmd.assert()
+        .failure()
+        .stderr(predicate::str::contains("Invalid Op"));
+
+    Ok(())
+}
+
+#[test]
+fn command_count() -> Result<(), Box<dyn std::error::Error>> {
+    let mut cmd = Command::cargo_bin("rgr")?;
+    let output = cmd
+        .arg("count")
+        .arg("tests/spanr/S288c.rg")
+        .arg("tests/spanr/S288c.rg")
+        .output()
+        .unwrap();
+    let stdout = String::from_utf8(output.stdout).unwrap();
+
+    assert_eq!(stdout.lines().count(), 6);
+    assert!(stdout.contains("I:1-100\t2"));
+    assert!(stdout.contains("21294-22075\t1"));
+
+    Ok(())
+}